BeliefPropagation.jl Documentation

Welcome to the documentation page.

Quick-Glimpse Tutorial

This tutorial just offered a quick glimpse on Julia's built-in documentation system, make sure to read the docs for more.

BeliefPropagation — Module

Belief propagation algorithm applied to the one-dimensional Ising model (Ising chain).

TBW

BeliefPropagation.get_cavity_parameter — Method

get_cavity_parameter(ω::T, β::T, h::T) where T <: Real

Get recursively the next cavity parameter in the chain.

The recursive relation between the cavity parameters is given by

\[\omega_{i-1}^{(i)} = h + \frac{1}{2\beta} \log{\left( \frac{ \cosh{[\beta(\omega_{i-2}^{(i-1)} + 1)]} }{ \cosh{[\beta(\omega_{i-2}^{(i-1)} - 1)]} } \right)}\]